group_a/requirements.txt
anthropic>=0.40.0
psycopg2-binary>=2.9.9
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
